Remove secrets from docker-compose.override.yml
This commit removes secrets from docker-compose.override.yml in favour of
storing these in 1Password and pulling them in at runtime via the use of
a .env file.

It includes a modified version of the script created by @Mhowell494 to make
the .env file available at the time of running the app and to ensure that
it is cleaned up after the app has exited.