Track Prosecution Concluded messages that fail processing
This commit adds tracking for messages from the Prosecution Concluded
queue which fail to be processed due to a number of possible validation
issues. These messages were previously dropped without being recorded
when they failed to be processed.

In order to start to understand the shape of the data being added to
the Prosecution Concluded queue so that possible remediations can be
introduced upstream and help to prevent these issues, all messages which
fail to be processed are added into a new `DeadLetterMessage` table as
part of this commit, which can then be inspected manually on a periodic
basis.

As part of this commit, stricter controls have also been added as to what
constitutes a valid Prosecution Concluded message, such that:

- the MAAT ID must be in a valid format
- the OU Code must not be empty and must match an existing OU Code
- there must be a Magistrates Court outcome when a Crown Court outcome
is specified

These validation controls have been front-loaded as much as possible to
ensure that processing fails fast if a message cannot be processed.